requirements.txt
# Core
pandas
numpy
scikit-learn
omegaconf
hydra-core

# Text processing
nltk

# Optional visualization / dev
tqdm
matplotlib
seaborn
